refactor: extract trade validators
