+++ packages/bridge-controller/src/bridge-controller.sse.test.ts
@@ -30,14 +30,15 @@ import {
   ETH_USDT_ADDRESS,
 } from './constants/bridge';
 import { ChainId, RequestStatus } from './types';
-import type { BridgeControllerMessenger, TxData } from './types';
+import type { BridgeControllerMessenger } from './types';
 import * as balanceUtils from './utils/balance';
 import { formatChainIdToDec } from './utils/caip-formatters';
 import * as featureFlagUtils from './utils/feature-flags';
 import * as fetchUtils from './utils/fetch';
+import { FeatureId } from './validators/feature-flags';
 import { QuoteStreamCompleteReason } from './validators/quote-stream-complete';
 import { TokenFeatureType } from './validators/token-feature';
-import { FeatureId } from './validators/feature-flags';
+import type { TxData } from './validators/trade';
 
 type RootMessenger = Messenger<
   MockAnyNamespace,
