+++ packages/bridge-controller/src/utils/metrics/properties.ts
@@ -7,8 +7,9 @@ import type {
   QuoteMetadata,
   QuoteRequest,
   QuoteResponseV1,
-  TxData,
 } from '../../types';
+import { FeatureId } from '../../validators/feature-flags';
+import type { TxData } from '../../validators/trade';
 import { getNativeAssetForChainId, isCrossChain } from '../bridge';
 import {
   formatAddressToAssetId,
@@ -22,7 +23,6 @@ import type {
   QuoteWarning,
   RequestParams,
 } from './types';
-import { FeatureId } from '../../validators/feature-flags';
 
 export const toInputChangedPropertyKey: Partial<
   Record<keyof QuoteRequest, InputKeys>
