+++ packages/bridge-controller/src/utils/trade-utils.test.ts
@@ -1,12 +1,16 @@
-import type { BitcoinTradeData, TronTradeData, TxData } from '../types';
+import type {
+  TxData,
+  BitcoinTradeData,
+  TronTradeData,
+  Trade,
+} from '../validators/trade';
 import {
-  extractTradeData,
   isEvmTxData,
   isBitcoinTrade,
   isStellarTrade,
   isTronTrade,
-} from './trade-utils';
-import type { Trade } from './trade-utils';
+} from '../validators/trade';
+import { extractTradeData } from './trade-utils';
 
 describe('Trade utils', () => {
   describe('isEvmTxData', () => {
